flow: revert anti-flood protection
Queries on busy systems do not reconnect when riemann is
restarted. Requesting the bot to reboot or sending a HUP to the query
does work and will force a reconnect.

Haven't been able to reproduce the behaviour so far on a test riemann
system.

Best guess is the bot has some high volume queries and XMPP throttling
limits draining the event queue. The result is either:

* flood protection kicks in and drops the control messages from gun

* the gun control messages are stuck in the queue while the evbot XMPP
  user waits for bandwidth to write to the XMPP channel

Since ejabberd throttles by user and the bot is responsive to other
messages and queries using other sources continue to work (e.g.,
containerized linux processes), try removing the flood protection from
the flow command.

The anti-flood mechanism could be fixed to match only %Runlet.Event{}
structs. The original commit message:

~~~
flow: add high water mark for the process mailbox

Under event floods, the bot may be throttled by the XMPP server. For
example, the default traffic shaper for ejabberd sets a limit of 1000
bytes/second.

If the bot is attempting to relay events at a rate of 10,000 bytes/second,
it will be throttled and messages will accumulate in the process
mailbox. The throttling will result in old messages being sent as
the bot attempts to catch up. Eventually the bot process may use up
available memory.

Flow control may also not be activated if the throttling limits
throughput to a rate below the flow control limits (default: 20
messages/100 seconds).

One fix is to increase the traffic shaper limits for the bot. Under
floods, the bot may still encounter throttling.

Set a high water mark for the mailbox. If the number of messages exceeds
this number (1000 by default), all messages are purged.

Problems:

* message types are not selectively purged

  Control (signal, ...) and event messages are both dropped.

* the check runs on every message
~~~