Runlet.State.table: open file read-only
table/2,3 retrieves all the rows from the state file. By default, dets
opens the file read-write. Reduce the opportunity for file corruption
by reducing the capability to read-only.

When the dets access mode is set to :read, the file is not created if it
does not exist. Match the error and return an empty list.

For other file errors (:eperm, :enospc, :emfile, ...), the calling
process will be crashed with a case clause error (previously, the
calling process was crashed with a match error).