flow: rate: remaining events in window
Convert the rate metric from events/second (float) to a count of the
remaining allotted events in the window (integer).

Previously the flow rate metric was the total number of events divided
by the runtime of the process.

The rate signals the noisiness when testing a runlet. In real usage a
runlet will be running for months so the rate always tends to ~0.

Alternatively the rate could be reset every window. Since events above
the threshold are dropped, the rate would have to be emitted at the
start of a new window for the previous window.

The rate metric now indicates the event threshold for the window. When
the rate is 0, further events are dropped.