fifo: named pipe as runlet input
Add a command to read the output of another runlet as input to a runlet:

    fifo foo
    query 'service = "bar"' > foo
    runtime > foo

Redirection between runlets is supported by using registered
processes. Using a PID also works but is useful only for testing because
the PIDs change between runlet invocations.

If the fifo does not exist, events are discarded:

* same behaviour as using a PID
* race condition between runlet startup on boot

  Runlet startup order is non-deterministic:

        query 'service = "bar"' > foo
        fifo foo